Added config for vendor-data-distribution-service
This services is now included in the Docker config. It can use a test
route, so that is added to the dispatcher's config. There is also some
config for the delta-notifier for it to work properly, and an added
config for mu-auth to allow SPARQL requests to be able to read the data
produced by the vendor-data-distribution-service.